Fix tests restructure
