+++ crates/types-ob-v3/tests/obv3_json_examples/endorsement_credential.json
