+++ crates/types-ob-v3/tests/obv3_json_examples/full_achievement_credential.json
