+++ crates/types-ob-v3/tests/obv3_json_examples/skill_assertion_credential_engine.json
