+++ crates/types-ob-v3/tests/tests/basic_achievement_credential.rs
@@ -1,10 +1,5 @@
 use crate::tests::assert_eq_json_value;
-use openbadges::{
-    achievement::{Achievement, AchievementBuilder, Criteria, CriteriaBuilder},
-    achievement_credential::{AchievementCredential, AchievementCredentialBuilder, AchievementCredentialType},
-    achievement_subject::{AchievementSubject, AchievementSubjectBuilder},
-    profile::{Profile, ProfileBuilder},
-};
+use types_ob_v3::prelude::*; 
 use std::fs::File;
 
 #[test]
