+++ crates/types-ob-v3/tests/tests/skill_assertion_case.rs
@@ -1,25 +1,16 @@
 use crate::tests::assert_eq_json_value;
-use openbadges::{
-    achievement::{AchievementBuilder, AchievementType, CriteriaBuilder},
-    achievement_credential::{
-        AchievementCredential, AchievementCredentialBuilder, AchievementCredentialType, CredentialSchemaBuilder,
-    },
-    achievement_subject::AchievementSubjectBuilder,
-    alignment::{AlignmentBuilder, AlignmentTargetType},
-    general::ImageBuilder,
-    profile::ProfileBuilder,
-};
 use std::{fs::File, str::FromStr};
+use types_ob_v3::prelude::*;
 
 #[test]
-fn skill_assertion_credential_engine() {
+fn skill_assertion_case() {
     // Testing if serialization and deserialization between the OBv3 examples and our rust code works as needed.
 
-    assert_eq_json_value::<AchievementCredential>("tests/obv3_json_examples/skill_assertion_credential_engine.json");
+    assert_eq_json_value::<AchievementCredential>("tests/obv3_json_examples/skill_assertion_case.json");
 
     // Next, the builders are tested against the OBv3 examples
 
-    let skill_assertion_credential_engine: AchievementCredential = AchievementCredentialBuilder::default()
+    let skill_assertion_case: AchievementCredential = AchievementCredentialBuilder::default()
     .context(vec![
         "https://www.w3.org/2018/credentials/v1",
 		"https://purl.imsglobal.org/spec/ob/v3p0/context-3.0.2.json",
@@ -27,23 +18,30 @@ fn skill_assertion_credential_engine() {
     ])
     .id("http://1edtech.edu/credentials/3732")
     .type_(AchievementCredentialType::from(vec!["VerifiableCredential", "OpenBadgeCredential"]))
-    .name("Solve and graph linear equations and inequalities")
+    .name("Robot Programming Skill Credential")
+    .description("A badge recognizing the development of skills in robot implementation, specifically the software".to_string())
     .credential_subject(
         AchievementSubjectBuilder::default()
         .id("did:example:ebfeb1f712ebc6f1c276e12ec21".to_string())
         .type_("AchievementSubject",)
         .achievement(
             AchievementBuilder::default()
-            .id("https://example.com/achievements/math/linear-1")
+            .id("https://example.com/achievements/robotics/robot-programming")
             .type_("Achievement")
+            .criteria(
+                CriteriaBuilder::default()
+                .narrative("Cite strong and thorough textual evidence to support analysis of what the text says explicitly as well as inferences drawn from the text, including determining where the text leaves matters uncertain".to_string())
+            )
+            .description("Analyze a sample text")
+            .name("Text analysis")
             .alignment(
                 vec![AlignmentBuilder::default()
                 .type_("Alignment")
-                .target_code("ce-6369c51f-4d86-4592-a761-8b32ae70a045".to_string())
-                .target_framework("Ivy Tech Community College of Indiana, MATH 135, FINITE MATH".to_string())
-                .target_name("Solve and graph linear equations and inequalities")
-                .target_type(AlignmentTargetType::from_str("ceasn:Competency").unwrap())
-                .target_url("https://credentialfinder.org/competency/ce-6369c51f-4d86-4592-a761-8b32ae70a045")
+                .target_description("Robot software is a set of commands and procedures robots use to respond to input and perform autonomous tasks.".to_string())
+                .target_name("Robot Programming")
+                .target_framework("Example Robotics Framework".to_string())
+                .target_type(AlignmentTargetType::from_str("CFItem").unwrap())
+                .target_url("https://robotics-competencies.example.com/competencies/robot-programming")
             ])
             .achievement_type(AchievementType::from_str("Competency").unwrap())
             .creator(
@@ -55,21 +53,28 @@ fn skill_assertion_credential_engine() {
                 .description("Example Industry Group is a consortium of luminaries who publish skills data for common usage.".to_string())
                 .email("info@exammple.com".to_string())
             )
-            .criteria({
+            .criteria(
                 CriteriaBuilder::default()
-                .narrative("Learners must demonstrate understanding of linear algebra and graphic representation of linear equations.".to_string())
-            })
-            .description("This achievement represents developing capability to solve and graph linear equations and inequalities")
-            .image({
+                .narrative("Learners must present source code showing the ability for a robot to accept manual or sensor input and perform conditional actions in response.".to_string())
+            )
+            .description("This achievement represents developing capability to develop software for robotic applications.".to_string())
+            .image(
                 ImageBuilder::default()
-                .id("https://example.com/achievements/math/linear-1/image")
+                .id("https://example.com/achievements/robotics/robot-programming/image")
                 .type_("Image")
-                .caption("A line, sloping upward optimistically".to_string())
-            })
-            .name("Linear equations and inequalities")
+                .caption("A robot filled with ones and zeroes representing its programming".to_string())
+            )
+            .name("Robot Programming")
         )
     )
-    .issuer({
+    .evidence(
+        vec![EvidenceBuilder::default()
+        .id("https://github.com/somebody/project".to_string())
+        .type_("Evidence")
+        .name("Final Project Code".to_string())
+        .description("The source code for the 'Beeper 1.0' robot project. It responds by saying 'beep' when the 'beep' button is pressed.".to_string())
+    ])
+    .issuer(
         ProfileBuilder::default()
         .id("https://1edtech.edu/issuers/565049")
         .type_("Profile")
@@ -77,14 +82,14 @@ fn skill_assertion_credential_engine() {
         .url("https://1edtech.edu".to_string())
         .phone("1-222-333-4444".to_string())
         .description("1EdTech University provides online degree programs.".to_string())
-        .image({
+        .image(
             ImageBuilder::default()
             .id("https://1edtech.edu/logo.png")
             .type_("Image")
             .caption("1EdTech University logo".to_string())
-        })
+        )
         .email("registrar@1edtech.edu".to_string())
-    })
+    )
     .issuance_date("2022-07-01T00:00:00Z")
     .credential_schema(
         vec![CredentialSchemaBuilder::default()
@@ -97,21 +102,19 @@ fn skill_assertion_credential_engine() {
 
     // Here we test the built struct against the struct deserialized from the example .json file.
 
-    let file =
-        File::open("tests/obv3_json_examples/skill_assertion_credential_engine.json").expect("Failed to open file");
-    let skill_assertion_cred_engine_from_file: AchievementCredential =
+    let file = File::open("tests/obv3_json_examples/skill_assertion_case.json").expect("Failed to open file");
+    let skill_assertion_case_from_file: AchievementCredential =
         serde_json::from_reader(&file).expect("Couldn't read from file");
 
-    assert_eq!(skill_assertion_credential_engine, skill_assertion_cred_engine_from_file);
+    assert_eq!(skill_assertion_case, skill_assertion_case_from_file);
 
     // Here we test the built struct converted to a json_value against the json_value deserialized from the example .json file
 
-    let file =
-        File::open("tests/obv3_json_examples/skill_assertion_credential_engine.json").expect("Failed to open file");
+    let file = File::open("tests/obv3_json_examples/skill_assertion_case.json").expect("Failed to open file");
     let json_value_from_file: serde_json::Value = serde_json::from_reader(file).expect("Couldn't read from file");
 
     assert_eq!(
-        serde_json::to_value(skill_assertion_credential_engine).unwrap(),
+        serde_json::to_value(skill_assertion_case).unwrap(),
         json_value_from_file
     );
 }
